baseapp: Move code from setters.go to options.go
This is because these setters are intended to be used as options for baseapp.
